% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metal.R
\name{theme_metal}
\alias{theme_metal}
\title{Metal Map}
\usage{
theme_metal(base_size = 11, base_family = "Roboto")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A map theme with a death metal vibe
}
