% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minima.R
\name{theme_minima}
\alias{theme_minima}
\title{Minima Moralia}
\usage{
theme_minima(base_size = 11, base_family = "Minion Pro")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme inspired by Adorno's critique of post-enlightenment capitalism
Clean, minimal, reveals the fractures of modern life
}
