% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rale.R
\name{theme_rale}
\alias{theme_rale}
\title{Ride the rails}
\usage{
theme_rale(base_size = 11, base_family = "Raleway")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
Theme using Raleway as the main font
Clean and minimal
}
